chore(frontend): move dotenv files to `frontend`
The `.env` file should be loaded to inject environment variables into the frontend packages.
These should be loaded regardless if running in devcontainers or on native machines.

This also changes the devcontainer to use a docker compose file.
This is because the `--env-file` argument that was being used is a required
argument, whereas in docker compose, the `env_file` key is optional.