Fix Violations of and Reenable `Lint/RedundantRequireStatement`
This one is pretty simple:

> Checks for unnecessary `require` statement.

Looks like the only unnecessary `require` statement we have is `require
'set'`, which was made reduntant in Ruby 3.2

Despite still being on Ruby 3.0, we are preparing for an eventual
upgrade to Ruby 3.3, and so have configured Rubocop to target that
version instead of the one we're actually on.

Speculatively opening this PR anyway, just to check what breaks.

- https://docs.rubocop.org/rubocop/cops_lint.html#lintredundantrequirestatement
- https://www.rubydoc.info/gems/rubocop/RuboCop/Cop/Lint/RedundantRequireStatement