fix(dsco): use css modules with automatic import
Previously, esbuild was configured to use the 'style' output which results in CSS being
injected into the head of the document. Since on SSR mode, `document` does not exist, swap over
to using CSS Modules which is supported by Next.js.

In this new method, CSS modules are generated by esbuild and automatically imported via the banner.