Docker-Compose-Based MySQL and Redis for Local Development (#62579)
* Adds basic docker-compose.yml for development environment.

* Adds DOCKER.md.

* Adds docker/developers path to the dependabot metadata.

* Update DOCKER.md

Co-authored-by: Stephen Liang <stephenliang@users.noreply.github.com>

* Allows for port to be specified in the database host.

This adds the ability to specify a port via typical ':' syntax to the db_endpoint_* configuration fields. This will allow you to more easily spin up secondary databases for secondary development environments. In particular, this enhances the Docker version of things my allowing the port to be specified via `MYSQL_NATIVE_PORT` such that you can spin it up alongside a native database more easily (and defer uninstalling one if you are migrating.)

This requies some small changes in our utility scripts for getting a console-based REPL for the database connections.

This also updates references to the database user in the Docker context to use `root` since that was what was used to create the databases.

* Makes the default mysql port 3306.

* Adds redis port configuration for natively run redis.

* Adds docker/developers path to the dependabot metadata.

* move docker-specific documentation into docker-specific directory

* add links to official documentation, extract some steps common to each environment, and slightly reorganize

* move daily reference details above first-time setup details

* add note to stop native mysql and redis services if they exist

* misc markdown formatting tweaks

more misc markdown formatting tweaks

* add macos-specific instructions for stopping existing services

---------

Co-authored-by: wilkie <david.wilkinson@code.org>
Co-authored-by: Stephen Liang <stephenliang@users.noreply.github.com>