fix(component-library): conditionally import css
The css bundler used in `esbuild-sass-plugin` will only produce an `index.css` if CSS is in use by an entrypoint.

However, some entrypoints do not use CSS and therefore no `index.css` is generated. As such, attempt to conditionally import CSS and if it does not exist, swallow the error.

Note: this solution is not ideal and should be revisited in the future as it performs an import at runtime. Ideally, no such import would exist in the first place if we know there is no CSS.